Http Upstream client
This pull request introduces the HTTP Upstream client. The main goal is to be
backward compatible and introduce an HTTP upstream client.

If the config has any HTTP/s server to connect to, the HttpClient struct will
handle them. Any WebSocket will use the existing upstream client code.

If one or more HTTP clients are configured, they route all requests, and
Websocket clients will be used for subscriptions.

If no HTTP clients are configured, the fallback behavior is used, and all
requests and subscriptions are routed through the WebSocket upstream client.

If no WebSocket upstream clients are configured, then subscriptions are not
enabled, only method requests.